requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cryptography==3.4.6
Django==3.0.5
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
pytz==2021.1
reportlab==3.5.63
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
